+++ packages/article-byline/__tests__/ios/article-byline.ios.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Article Byline test on iOS", () => {
+  shared();
+});