+++ packages/article-byline/__tests__/web/article-byline.web.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Article Byline test on web", () => {
+  shared();
+});