+++ packages/utils/__tests__/jest.config.js
@@ -1,0 +1,6 @@
+const jestConfigurator = require("@times-components/jest-configurator").default;
+
+module.exports = jestConfigurator(null, __dirname, [
+  "graphql.js",
+  "fetch-gql-schema.js"
+]);