+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.4.0"></a>
+# [0.4.0](https://github.com/newsuk/times-components/compare/@times-components/utils@0.3.0...@times-components/utils@0.4.0) (2018-02-15)
+
+
+### Features
+
+* allow platform independence in Jest Configurator ([#646](https://github.com/newsuk/times-components/issues/646)) ([d83b7b2](https://github.com/newsuk/times-components/commit/d83b7b2))
+
+
+
+
 <a name="0.3.1"></a>
 ## [0.3.1](https://github.com/newsuk/times-components/compare/@times-components/utils@0.3.0...@times-components/utils@0.3.1) (2018-02-14)